Тупко #29: Windows callback - third step
