--- openssl/pbauto_heartbeat_callback_posix.c
@@ -1,1 +1,0 @@
-../posix/pbauto_heartbeat_callback_posix.c
