+++ posix/posix.mk
@@ -131,8 +131,8 @@ CALLBACK_INTF_OBJFILES += pubnub_parse_ipv6_addr.o
 endif
 
 ifeq ($(USE_AUTO_HEARTBEAT), 1)
-CALLBACK_INTF_SOURCEFILES += ../posix/pbauto_heartbeat_callback_posix.c
-CALLBACK_INTF_OBJFILES += pbauto_heartbeat_callback_posix.o
+CALLBACK_INTF_SOURCEFILES += ../core/pbauto_heartbeat_callback.c ../posix/pbauto_heartbeat_init_posix.c
+CALLBACK_INTF_OBJFILES += pbauto_heartbeat_callback.o pbauto_heartbeat_init_posix.o
 endif
 
 CFLAGS_CALLBACK = -D PUBNUB_USE_IPV6=$(USE_IPV6) -D PUBNUB_SET_DNS_SERVERS=$(USE_DNS_SERVERS)
